// src/app/types/modes.ts
export type Mode = 'AA-light' | 'AAA-light' | 'AA-dark' | 'AAA-dark';

export type SurfaceStyle = 
  | 'light-tonal' 
  | 'colorful-tonal' 
  | 'dark-tonal' 
  | 'light-professional' 
  | 'grey-professional' 
  | 'dark-professional' 
  | 'colorful-professional';

export type ButtonShape = 'gently-rounded' | 'amply-rounded' | 'boldly-rounded' | 'square';
export type ComponentEffect = 'none' | 'bevel' | 'ridged';
export type HotlinkStyle = 'tonal' | 'blue';